pyproject.toml
[project]
name = "cognibench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.72.0",
    "python-dotenv>=1.1.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0", # [standard] includes performance extras
    "pytest>=8.3.5",
    "pyyaml>=6.0", # Added for loading YAML config files
    "tqdm>=4.66.0", # Added for progress bars
    "watchdog>=6.0.0",
    "antlr4-python3-runtime>=4.13.2",
    "sympy>=1.13.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0", # Added pytest for testing
    "ruff>=0.4.0",   # Added ruff for linting and formatting
]

[tool.setuptools]
packages = ["core", "api"]
